+++ src/build.go
@@ -17,6 +17,7 @@ import (
 	"github.com/urfave/cli/v2"
 	"github.com/vyPal/CaffeineC/lib/compiler"
 	"github.com/vyPal/CaffeineC/lib/parser"
+	"github.com/vyPal/CaffeineC/lib/project"
 )
 
 func init() {
@@ -128,7 +129,7 @@ func build(c *cli.Context) error {
 		}
 	}
 
-	var conf CfConf
+	var conf project.CfConf
 
 	f := c.Args().First()
 	if f == "" {
@@ -136,7 +137,7 @@ func build(c *cli.Context) error {
 		if c.String("config") != "" {
 			confPath = c.String("config")
 		}
-		conf, err = GetCfConf(confPath)
+		conf, err = project.GetCfConf(confPath)
 		if err != nil {
 			return err
 		}
@@ -151,7 +152,7 @@ func build(c *cli.Context) error {
 	if os.IsNotExist(err) {
 		return cli.Exit(color.RedString("File does not exist: %s", f), 1)
 	} else if inf.IsDir() {
-		conf, err := GetCfConf(f)
+		conf, err := project.GetCfConf(f)
 		if err != nil {
 			return err
 		}
