+++ src/projects.go
@@ -1,23 +1,20 @@
 package main
 
 import (
-	"bufio"
-	"encoding/gob"
 	"fmt"
-	"log"
 	"net/url"
 	"os"
 	"path"
 	"path/filepath"
-	"runtime"
 	"strings"
 
 	"github.com/fatih/color"
 	"github.com/go-git/go-git/v5"
 	"github.com/go-git/go-git/v5/plumbing"
 	"github.com/urfave/cli/v2"
+	"github.com/vyPal/CaffeineC/lib/cache"
+	"github.com/vyPal/CaffeineC/lib/project"
 	"github.com/vyPal/CaffeineC/util"
-	"gopkg.in/yaml.v3"
 )
 
 func init() {
@@ -86,98 +83,41 @@ func init() {
 	})
 }
 
-type CfConf struct {
-	Name         string             `yaml:"name"`
-	Description  string             `yaml:"description"`
-	Version      string             `yaml:"version"`
-	Main         string             `yaml:"main"`
-	Dependencies []CFConfDependency `yaml:"dependencies"`
-	Author       string             `yaml:"author"`
-	License      string             `yaml:"license"`
-}
-
-type CFConfDependency struct {
-	Package    string `yaml:"package"`
-	Version    string `yaml:"version"`
-	Identifier string `yaml:"identifier"`
-}
-
-func (c *CfConf) CreateDefault() {
-	c.Name = "NewProject"
-	c.Description = "A new CaffeineC project"
-	c.Version = "1.0.0"
-	c.Main = "src/main.cffc"
-	c.Author = "Anonymous"
-	c.License = "MIT"
-}
-
-func (c *CfConf) Save(filepath string, overwrite bool) error {
-	if _, err := os.Stat(filepath); !os.IsNotExist(err) {
-		if overwrite || promptYN(filepath+" already exists. Overwrite?", false) {
-			os.Remove(filepath)
-		} else {
-			return nil
-		}
-	}
-
-	_, err := os.Create(filepath)
-	if err != nil {
-		return err
-	}
-
-	file, err := os.OpenFile(filepath, os.O_WRONLY, 0644)
-	if err != nil {
-		return err
-	}
-
-	yml, err := yaml.Marshal(c)
-	if err != nil {
-		return err
-	}
-
-	_, err = file.Write(yml)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
 func install(c *cli.Context) error {
 	liburl := c.Args().First()
-	conf, err := GetCfConf("")
+	conf, err := project.GetCfConf("")
 	if err != nil {
 		return err
 	}
 
-	cache := PackageCache{}
-	err = cache.Init()
+	pcache := cache.PackageCache{}
+	err = pcache.Init()
 	if err != nil {
 		return err
 	}
 
 	if liburl == "" {
 		for _, dep := range conf.Dependencies {
-			pkg, err := cache.GetPackage(dep.Package, dep.Version, dep.Identifier)
+			pkg, err := pcache.GetPackage(dep.Package, dep.Version, dep.Identifier)
 			if err != nil {
 				return err
 			}
 
-			if pkg == (Package{}) {
+			if pkg == (cache.Package{}) {
 				color.Green("Package not found locally, cloning...")
-				conf, _, _, err = InstallLibrary(cache, dep.Identifier)
+				conf, _, _, err = InstallLibrary(pcache, dep.Identifier)
 				if err != nil {
 					return err
 				}
 			} else {
-				conf, err = GetCfConf(pkg.Path)
+				conf, err = project.GetCfConf(pkg.Path)
 				if err != nil {
 					return err
 				}
 			}
 		}
 	} else {
-		err = cache.CacheScan(true)
+		err = pcache.CacheScan(true)
 		if err != nil {
 			return err
 		}
@@ -187,28 +127,28 @@ func install(c *cli.Context) error {
 			return err
 		}
 
-		pkg, err := cache.GetPackage("", "", filepath.Join(strings.TrimPrefix(liburl, "https://"), version))
+		pkg, err := pcache.GetPackage("", "", filepath.Join(strings.TrimPrefix(liburl, "https://"), version))
 		if err != nil {
 			return err
 		}
 
 		var ident, ver string
-		if pkg == (Package{}) {
+		if pkg == (cache.Package{}) {
 			color.Green("Package not found locally, cloning...")
-			conf, ident, ver, err = InstallLibrary(cache, liburl)
+			conf, ident, ver, err = InstallLibrary(pcache, liburl)
 			if err != nil {
 				return err
 			}
 		} else {
 			ident = pkg.Identifier
 			ver = pkg.Version
-			conf, err = GetCfConf(pkg.Path)
+			conf, err = project.GetCfConf(pkg.Path)
 			if err != nil {
 				return err
 			}
 		}
 
-		dep := CFConfDependency{
+		dep := project.CFConfDependency{
 			Package:    conf.Name,
 			Version:    ver,
 			Identifier: ident,
@@ -262,17 +202,17 @@ func PrepUrl(liburl string) (u, ver string, e error) {
 	return liburl, version, nil
 }
 
-func InstallLibrary(cache PackageCache, liburl string) (conf CfConf, ident, ver string, e error) {
+func InstallLibrary(pcache cache.PackageCache, liburl string) (conf project.CfConf, ident, ver string, e error) {
 	liburl, version, err := PrepUrl(liburl)
 	if err != nil {
-		return CfConf{}, "", "", err
+		return project.CfConf{}, "", "", err
 	}
 
 	// Create a directory in the cache's BaseDir
-	installDir := filepath.Join(cache.BaseDir, strings.TrimPrefix(liburl, "https://"), version)
+	installDir := filepath.Join(pcache.BaseDir, strings.TrimPrefix(liburl, "https://"), version)
 	err = os.MkdirAll(installDir, 0700)
 	if err != nil {
-		return CfConf{}, "", "", err
+		return project.CfConf{}, "", "", err
 	}
 
 	// Clone the repository to the install directory
@@ -283,25 +223,25 @@ func InstallLibrary(cache PackageCache, liburl string) (conf CfConf, ident, ver 
 		ReferenceName: plumbing.NewBranchReferenceName(version),
 	})
 	if err != nil {
-		return CfConf{}, "", "", err
+		return project.CfConf{}, "", "", err
 	}
 
-	pkg := Package{
+	pkg := cache.Package{
 		Name:       conf.Name,
 		Version:    conf.Version,
 		Identifier: strings.TrimPrefix(liburl, "https://"),
 		Path:       installDir,
 	}
-	cache.PkgList = append(cache.PkgList, pkg)
-	err = cache.CacheSave()
+	pcache.PkgList = append(pcache.PkgList, pkg)
+	err = pcache.CacheSave()
 	if err != nil {
-		return CfConf{}, "", "", err
+		return project.CfConf{}, "", "", err
 	}
 
 	// Get the configuration file from the cloned repository
-	conf, err = GetCfConf(installDir)
+	conf, err = project.GetCfConf(installDir)
 	if err != nil {
-		return CfConf{}, "", "", err
+		return project.CfConf{}, "", "", err
 	}
 
 	return conf, strings.TrimPrefix(liburl, "https://"), version, nil
@@ -309,31 +249,31 @@ func InstallLibrary(cache PackageCache, liburl string) (conf CfConf, ident, ver 
 
 func libInfo(c *cli.Context) error {
 	liburl := c.Args().First()
-	var conf CfConf
+	var conf project.CfConf
 	var err error
 	if liburl == "" {
-		conf, err = GetCfConf("")
+		conf, err = project.GetCfConf("")
 		if err != nil {
 			return err
 		}
 	} else {
-		cache := PackageCache{}
-		err = cache.Init()
+		pcache := cache.PackageCache{}
+		err = pcache.Init()
 		if err != nil {
 			return err
 		}
 
-		err = cache.CacheScan(true)
+		err = pcache.CacheScan(true)
 		if err != nil {
 			return err
 		}
 
-		pkg, err := cache.GetPackage("", "", liburl)
+		pkg, err := pcache.GetPackage("", "", liburl)
 		if err != nil {
 			return err
 		}
 
-		if pkg == (Package{}) {
+		if pkg == (cache.Package{}) {
 			color.Green("Package not found locally, cloning...")
 			liburl, version, err := PrepUrl(liburl)
 			if err != nil {
@@ -359,12 +299,12 @@ func libInfo(c *cli.Context) error {
 			}
 
 			// Get the configuration file from the cloned repository
-			conf, err = GetCfConf(tempDir)
+			conf, err = project.GetCfConf(tempDir)
 			if err != nil {
 				return err
 			}
 		} else {
-			conf, err = GetCfConf(pkg.Path)
+			conf, err = project.GetCfConf(pkg.Path)
 			if err != nil {
 				return err
 			}
@@ -385,177 +325,6 @@ func libInfo(c *cli.Context) error {
 	return nil
 }
 
-type PackageCache struct {
-	BaseDir string
-	RootDir string
-	PkgList []Package
-}
-
-type Package struct {
-	Name       string
-	Version    string
-	Identifier string
-	Path       string
-}
-
-func (p *PackageCache) Init() error {
-	homeDir, err := os.UserHomeDir()
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	libDir := path.Join(homeDir, ".local", "lib", "CaffeineC")
-	if runtime.GOOS == "windows" {
-		libDir = path.Join(homeDir, "AppData", "Local", "Programs", "CaffeineC")
-	}
-
-	err = os.MkdirAll(libDir, 0700)
-	if err != nil {
-		return err
-	}
-
-	cacheDir := path.Join(libDir, "packages")
-	err = os.Mkdir(cacheDir, 0700)
-	if err != nil && !os.IsExist(err) {
-		return err
-	}
-
-	p.RootDir = libDir
-	p.BaseDir = cacheDir
-	p.PkgList = make([]Package, 0)
-
-	return nil
-}
-
-func (p *PackageCache) DeepCacheScan() error {
-	err := filepath.WalkDir(p.BaseDir, func(path string, d os.DirEntry, err error) error {
-		if err != nil {
-			return nil
-		}
-
-		if d.IsDir() {
-			conf, err := GetCfConf(path)
-			if err != nil {
-				if os.IsNotExist(err) {
-					return nil
-				} else {
-					return err
-				}
-			}
-
-			identifier := strings.TrimPrefix(path, p.BaseDir)
-			identifier = strings.TrimPrefix(identifier, "/")
-
-			pkg := Package{
-				Name:       conf.Name,
-				Version:    conf.Version,
-				Identifier: identifier,
-				Path:       path,
-			}
-
-			p.PkgList = append(p.PkgList, pkg)
-			return filepath.SkipDir
-		}
-
-		return nil
-	})
-
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (p *PackageCache) CacheScan(deepOnFail bool) error {
-	cacheFile, err := os.Open(path.Join(p.BaseDir, "cache.bin"))
-	if err != nil {
-		if os.IsNotExist(err) {
-			if deepOnFail {
-				fmt.Println("Cache file not found, performing deep scan...")
-				err := p.DeepCacheScan()
-				if err != nil {
-					return err
-				}
-
-				err = p.CacheSave()
-				if err != nil {
-					return err
-				}
-				return nil
-			} else {
-				return err
-			}
-		} else {
-			return err
-		}
-	}
-
-	decoder := gob.NewDecoder(cacheFile)
-	err = decoder.Decode(&p.PkgList)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (p *PackageCache) CacheSave() error {
-	cacheFile, err := os.Create(path.Join(p.BaseDir, "cache.bin"))
-	if err != nil {
-		return err
-	}
-
-	encoder := gob.NewEncoder(cacheFile)
-	err = encoder.Encode(p.PkgList)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (p *PackageCache) GetPackage(name, version, identifier string) (Package, error) {
-	for _, pkg := range p.PkgList {
-		fmt.Printf("Name: %s, Version: %s, Ident: %s, SearchIdent: %s\n", pkg.Name, pkg.Version, pkg.Identifier, identifier)
-		if (pkg.Name == name || name == "") && pkg.Identifier == identifier {
-			if version == "" || version == "*" || version == pkg.Version {
-				return pkg, nil
-			}
-			continue
-		}
-	}
-
-	return Package{}, nil
-}
-
-func (p *PackageCache) HasPackage(name, version, identifier string) (bool, error) {
-	for _, pkg := range p.PkgList {
-		if pkg.Name == name && pkg.Identifier == identifier {
-			if version == "" || version == "*" {
-				return true, nil
-			}
-
-			sver, err := util.Parse(pkg.Version)
-			if err != nil {
-				return false, err
-			}
-
-			sat, err := sver.Satisfies(version)
-			if err != nil {
-				return false, err
-			}
-
-			if sat {
-				return true, nil
-			}
-			continue
-		}
-	}
-
-	return false, nil
-}
-
 func initProject(c *cli.Context) error {
 	rootDir := c.Args().First()
 	if rootDir == "" {
@@ -571,7 +340,7 @@ func initProject(c *cli.Context) error {
 		}
 
 		if len(files) > 0 {
-			if !promptYN("The directory is not empty, continue?", false) {
+			if !util.PromptYN("The directory is not empty, continue?", false) {
 				return nil
 			}
 		}
@@ -612,8 +381,8 @@ func initProject(c *cli.Context) error {
 		fmt.Println("Created file:", path.Join(rootDir, "src", "main.cffc"))
 	}
 
-	if promptYN("Use default configuration?", false) {
-		conf := CfConf{}
+	if util.PromptYN("Use default configuration?", false) {
+		conf := project.CfConf{}
 		conf.CreateDefault()
 
 		err := conf.Save(path.Join(rootDir, "cfconf.yaml"), false)
@@ -623,14 +392,14 @@ func initProject(c *cli.Context) error {
 
 		fmt.Println("Created file:", path.Join(rootDir, "cfconfig.yaml"))
 	} else {
-		conf := CfConf{}
+		conf := project.CfConf{}
 
-		conf.Name = promptString("Project name", "NewProject")
-		conf.Description = promptString("Project description", "A new CaffeineC project")
-		conf.Version = promptString("Project version", "1.0.0")
-		conf.Main = promptString("Main file", "src/main.cffc")
-		conf.Author = promptString("Author", "Anonymous")
-		conf.License = promptString("License", "MIT")
+		conf.Name = util.PromptString("Project name", "NewProject")
+		conf.Description = util.PromptString("Project description", "A new CaffeineC project")
+		conf.Version = util.PromptString("Project version", "1.0.0")
+		conf.Main = util.PromptString("Main file", "src/main.cffc")
+		conf.Author = util.PromptString("Author", "Anonymous")
+		conf.License = util.PromptString("License", "MIT")
 
 		err := conf.Save(path.Join(rootDir, "cfconf.yaml"), false)
 		if err != nil {
@@ -646,62 +415,4 @@ func initProject(c *cli.Context) error {
 	fmt.Println("----------------------------------------")
 
 	return nil
-}
-
-func GetCfConf(dir string) (CfConf, error) {
-	var conf CfConf
-
-	file, err := os.Open(path.Join(dir, "cfconf.yaml"))
-	if err != nil {
-		return CfConf{}, err
-	}
-
-	decoder := yaml.NewDecoder(file)
-	err = decoder.Decode(&conf)
-	if err != nil {
-		return CfConf{}, err
-	}
-
-	return conf, nil
-}
-
-func promptYN(prompt string, def bool) bool {
-	reader := bufio.NewReader(os.Stdin)
-
-	if def {
-		fmt.Printf("%s (Y/n): ", prompt)
-	} else {
-		fmt.Printf("%s (y/N): ", prompt)
-	}
-
-	response, err := reader.ReadString('\n')
-	if err != nil {
-		panic(err)
-	}
-
-	response = strings.TrimSpace(response)
-
-	if response == "" {
-		return def
-	}
-
-	return strings.ToLower(response) == "y"
-}
-
-func promptString(prompt string, def string) string {
-	reader := bufio.NewReader(os.Stdin)
-	fmt.Printf("%s (%s): ", prompt, def)
-
-	response, err := reader.ReadString('\n')
-	if err != nil {
-		panic(err)
-	}
-
-	response = strings.TrimSpace(response)
-
-	if response == "" {
-		return def
-	}
-
-	return response
 }
